requirements.txt
uvicorn
fastapi
pydantic
requests
requests
numpy
torch
pandas
opencv_python
scikit_image
matplotlib
dlib
mtcnn
torchvision
visdom
pillow
scikit_learn
